Add explicit cast

+++ apps/super_simplify/expr_util.cpp
@@ -346,7 +346,7 @@ void all_possible_exprs_that_compute_associative_op_helper(const Expr &e,
     }
 
     vector<Expr> terms = unpack_binary_op<Op>(e);
-    for (size_t i = 1; i < (1 << terms.size()) - 1; i++) {
+    for (size_t i = 1; i < (size_t)((1 << terms.size()) - 1); i++) {
         vector<Expr> left, right;
         for (size_t j = 0; j < terms.size(); j++) {
             if (i & (1 << j)) {
